fixed comments thumbs up/down state incorrect error
