+++ internal/core/comments.go
@@ -25,7 +25,7 @@ type CommentService interface {
 	GetCommentReplyByID(id int64) (*CommentReply, error)
 	GetCommentContentsByIDs(ids []int64) ([]*CommentContent, error)
 	GetCommentRepliesByID(ids []int64) ([]*CommentReplyFormated, error)
-	GetCommentThumbsMap(tweetId int64) (cs.CommentThumbsMap, cs.CommentThumbsMap, error)
+	GetCommentThumbsMap(userId int64, tweetId int64) (cs.CommentThumbsMap, cs.CommentThumbsMap, error)
 }
 
 // CommentManageService 评论管理服务
