+++ internal/model/web/pub.go
@@ -6,7 +6,6 @@ package web
 
 import (
 	"github.com/rocboss/paopao-ce/internal/core"
-	"github.com/rocboss/paopao-ce/internal/servants/base"
 	"github.com/rocboss/paopao-ce/pkg/version"
 )
 
@@ -15,15 +14,6 @@ type TweetDetailReq struct {
 }
 
 type TweetDetailResp core.PostFormated
-
-type TweetCommentsReq struct {
-	TweetId      int64  `form:"id"`
-	SortStrategy string `form:"sort_strategy"`
-	Page         int    `form:"-"`
-	PageSize     int    `form:"-"`
-}
-
-type TweetCommentsResp base.PageResp
 
 type GetCaptchaResp struct {
 	Id      string `json:"id"`
