+++ mirc/web/v1/loose.go
@@ -26,4 +26,7 @@ type Loose struct {
 
 	// TopicList 获取话题列表
 	TopicList func(Get, web.TopicListReq) web.TopicListResp `mir:"/tags"`
+
+	// TweetComments 获取动态评论
+	TweetComments func(Get, web.TweetCommentsReq) web.TweetCommentsResp `mir:"/post/comments"`
 }
