+++ mirc/web/v1/pub.go
@@ -31,7 +31,4 @@ type Pub struct {
 
 	// TweetDetail 获取动态详情
 	TweetDetail func(Get, web.TweetDetailReq) web.TweetDetailResp `mir:"/post"`
-
-	// TweetComments 获取动态评论
-	TweetComments func(Get, web.TweetCommentsReq) web.TweetCommentsResp `mir:"/post/comments"`
 }
